factored out methods into seperate lib directory
